moved `_migrate_icon_fa4_value` to upgrades
