+++ core/upgrade/upgrades.php
@@ -1,6 +1,7 @@
 <?php
 namespace Elementor\Core\Upgrade;
 
+use Elementor\Icons_Manager;
 use Elementor\Plugin;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -507,6 +508,35 @@ class Upgrades {
 	}
 
 	/**
+	 * migrate Icon control string value to Icons control array value
+	 *
+	 * @param array $element
+	 * @param array $args
+	 *
+	 * @return mixed
+	 */
+	public static function _migrate_icon_fa4_value( $element, $args ) {
+		$widget_id = $args['widget_id'];
+
+		if ( empty( $element['widgetType'] ) || $widget_id !== $element['widgetType'] ) {
+			return $element;
+		}
+		foreach ( $args['control_ids'] as $old_name => $new_name ) {
+			// exit if new value exists
+			if ( isset( $element['settings'][ $new_name ] ) ) {
+				continue;
+			}
+			// exit if no value to migrate
+			if ( empty( $element['settings'][ $old_name ] ) && ! isset( $element['settings'][ $new_name ] ) ) {
+				continue;
+			}
+			$element['settings'][ $new_name ] = Icons_Manager::fa4_to_fa5_value_migration( $element['settings'][ $old_name ] );
+			$args['do_update'] = true;
+		}
+		return $element;
+	}
+
+	/**
 	 * Set FontAwesome 5 value Migration on for button widget
 	 *
 	 * @param Updater $updater
@@ -514,7 +544,7 @@ class Upgrades {
 	public static function _v_2_6_6_fa4_migration_button( $updater ) {
 		$changes = [
 			[
-				'callback' => [ 'Elementor\Core\Upgrade\Upgrade_Utils', '_migrate_icon_fa4_value' ],
+				'callback' => [ 'Elementor\Core\Upgrade\Upgrades', '_migrate_icon_fa4_value' ],
 				'control_ids' => [
 					'icon' => 'selected_icon',
 				],
